Fix Table requests
